changed name of file to useApplicationData
